Refactored project to accommodate Graves support.
