+++ src/main/java/de/snap20lp/offlineplayers/depends/APIManager.java
@@ -1,7 +1,8 @@
 package de.snap20lp.offlineplayers.depends;
 
-import de.snap20lp.offlineplayers.DepartedDepotIntegration;
+import de.snap20lp.offlineplayers.depends.integrations.DepartedDepotIntegration;
 import de.snap20lp.offlineplayers.OfflinePlayers;
+import de.snap20lp.offlineplayers.depends.integrations.RanullGravesIntegration;
 import org.bukkit.Bukkit;
 
 import javax.annotation.Nullable;
@@ -73,11 +74,15 @@ public class APIManager {
             multiverseInventoriesFacade = null;
             OfflinePlayers.getInstance().getLogger().log(Level.WARNING, "Multiverse Inventories not found. Plugin will run normally.");
         }
-        // todo: Update to match the APIManager pattern.
         try {
             Bukkit.getPluginManager().registerEvents(new DepartedDepotIntegration(), OfflinePlayers.getInstance());
-        } catch (NoClassDefFoundError error) {
-            OfflinePlayers.getInstance().getLogger().log(Level.WARNING, "Departed Depots not found. Plugin will run normally.");
+        } catch (NoClassDefFoundError error1) {
+            OfflinePlayers.getInstance().getLogger().log(Level.WARNING, "Departed Depots not found. Attempting a graves fallback: Graves by Ranull.");
+            try {
+                Bukkit.getPluginManager().registerEvents(new RanullGravesIntegration(), OfflinePlayers.getInstance());
+            } catch (NoClassDefFoundError error2) {
+                OfflinePlayers.getInstance().getLogger().log(Level.WARNING, "Graves by Ranull not found. Plugin will run normally without graves.");
+            }
         }
     }
